+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/commerce_product.ftl
@@ -3,7 +3,7 @@
 
 	commerceCatalogModel = commerceDataFactory.newCommerceCatalogModel(commerceCurrencyModel)
 
-	commerceCatalogGroupModel = dataFactory.newCommerceCatalogGroupModel(commerceCatalogModel, classNameDataFactory.getClassNameId("CommerceCatalog"))
+	commerceCatalogGroupModel = userDataFactory.newCommerceCatalogGroupModel(commerceCatalogModel, classNameDataFactory.getClassNameId("CommerceCatalog"))
 	commerceChannelModel = commerceDataFactory.newCommerceChannelModel(commerceCurrencyModel)
 	cpTaxCategoryModel = commerceDataFactory.newCPTaxCategoryModel()
 />
@@ -42,4 +42,4 @@ ${dataFactory.toInsertSQL(cpTaxCategoryModel)}
 
 <@insertGroup _groupModel=commerceCatalogGroupModel />
 
-<@insertGroup _groupModel=dataFactory.newCommerceChannelGroupModel(commerceChannelModel, classNameDataFactory.getClassNameId("CommerceChannel")) />
+<@insertGroup _groupModel=userDataFactory.newCommerceChannelGroupModel(commerceChannelModel, classNameDataFactory.getClassNameId("CommerceChannel")) />
